fix: keep hook policy registry internal
